rtl8723du: Remove CONFIG_RFKILL_POLL, RTW_HALMAC, CONFIG_MCC_MODE, CONFIG_XTAL_26M, CONFIG_WIFI_TEST. CONFIG_SIGNAL_SCALE_MAPPING, CONFIG_APPEND_VENDOR_IE_ENABLE, CONFIG_RTW_IPCAM_APPLICATION, CONFIG_WIFI_MONITOR, CONFIG_MULTIDRV, CONFIG_AUTOCFG_CP, CONFIG_80211W, CONFIG_RF4CE_COEXIST
CONFIG_RFKILL_POLL is not selected.
RTW_HALMAC is not selected.
CONFIG_MCC_MODE is not selected.
CONFIG_XTAL_26M is not selected.
CONFIG_WIFI_TEST is not selected.
CONFIG_SIGNAL_SCALE_MAPPING is not selected.
CONFIG_APPEND_VENDOR_IE_ENABLE is not selected.
CONFIG_RTW_IPCAM_APPLICATION is not selected.
CONFIG_WIFI_MONITOR is selected.
CONFIG_MULTIDRV is not selected.
CONFIG_AUTOCFG_CP is not selected.
CONFIG_80211W is not selected. CONFIG_IEEE80211W has not been removed and may be implemented later.
CONFIG_RF4CE_COEXIST is not selected.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>