rtl8723du: Fix a number of endian issues
There are two places that still show errors. I am not concerned with
core/rtw_br_ext.c as it could not have worked on a little-endian system.
A couple of the Sparse warnings related to entries in the _io_ops
are still being investigated, but I do not think they will affect
execution.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>